remove compare drawer as registered component
