--- apps/web/vibes/soul/docs/compare-drawer.mdx
@@ -1,4 +1,0 @@
----
-title: Compare Drawer
-preview: compare-drawer-example
----
