+++ apps/web/vibes/soul/examples.ts
@@ -60,13 +60,6 @@ export const examples = [
     component: lazy(() => import('./examples/checkbox')),
   },
   {
-    name: 'compare-drawer-example',
-    dependencies: [],
-    registryDependencies: ['compare-drawer'],
-    files: ['examples/compare-drawer.tsx'],
-    component: lazy(() => import('./examples/compare-drawer')),
-  },
-  {
     name: 'discount-example',
     dependencies: [],
     registryDependencies: ['discount'],
